pyproject.toml
[project]
name = "ingregenius"
version = "0.1.0"
description = ""
authors = [
    {name = "Aayush"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab (>=4.4.7,<5.0.0)",
    "roboflow (>=1.2.9,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "streamlit (>=1.50.0,<2.0.0)",
    "ultralytics (>=8.3.204,<9.0.0)",
    "openai (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry]
packages = [{ include = "ingregenius", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
